Website/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2",
    "djangorestframework>=3.16.0",
    "pillow>=11.2.1",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.0",
    "sympy>=1.14.0",
]
